Add a prototype of expression inlining for icxx
This implementation is inefficient, but I can't think of any other way
to do it at the moment. In particular, the problem is that we really want
the quoting behavior from the closure (and there's no good way to implement
this without using one), but I can't see a way to get the LLVM code of the
closure through to the staged function in an efficient manner (type inference
would have to be able to compute this in order for it to be useful/more efficient
than this). Maybe as the julia type system is redesigned we can take another look.