Refactor everything to allow for multiple active Clang instances
This is useful when dealing with code that can't live in the same compilation unit,
e.g. multiple versions of LLVM. It is also a better practice to avoid global state.
On the C++ side, this goal is achieved by factoring Clang's objects into a single
struct (CxxInstance) that is passed around everywhere. On the Julia side, we also
need to pass around this struct, however, an effort is made to preserve simplicity,
when only one compiler instance is in use. Functions such as addHeaderDir, cxxinclude,
etc. take an additional parameter that indicates which compiler to use, but fall back
to the first (default) compiler instance. Macros have the same behavior, however, the
compiler instance to use can be globally overriden for a scope, by setting the
__current_compiler__ variable. New compiler instances can be created using the
`new_clang_instance`. E.g., to use a new compiler instance just for once specific thing,
one might do
```
let __current_compiler__ = new_clang_instance()
    cxx"""
    #include <...>
    """
    ...
end
```
Note that a new compiler instance will have to parse every file again, and will use
memory/compute doing so.