Implement #103
This turned out to be quite tricky, because contrary to my expectation,
lambda's are not generally convertible to function pointers. To expose
them to julia, we need to pretend we're a templated function and conjur
up implementations after the fact. This is still a prototype, but it
seems to work.