Translate anonymous enums as integers. Fixes #118
I don't believe there is any case in which one would need to recreate a
value whose type is given by the anonymous enum, making the declared
names essentially glorified integer constants. Since we don't have a name,
we can't declare them as a CppEnum anyway so we might as well treat them
as such.