Set Target Options correctly
Words can not describe the ammount of pain this caused. Suffice it to say
that libstdc++'s shared_ptr layout depends on which intrinsics the compiler
declares to have an without this, clang declares to have none, while any
library built for anything better than an i486, declares at least some.
And there goes a week of my life