Transition to NTuples for CppValues
this has long been the plan and the previous solution was a stopgap.
In order to make this work, I decided to pull the CVR qualifiers out of CppValues
into CxxQualType and have CppPtrs no longer longer refer to the CppValue directly,
but go through a CxxQualType instead. The reason for this is that the `N` field
of the CppValue is frequently unbound, so we can only have it at the top type level
without making that type nearly useless for dispatch.

Additionally, this gives us the ability to add finalizers to CppValues, so enable
that to automatically destruct C++ objects with non-trivial destructors.