support
"false">http://blog.bitbucket.org/?p=2473</guid>
<description>
<![CDATA[
Hi, I&#8217;m Sean, the newest member on the Bitbucket team. I currently work on Mercurial but come from a scientific background with the PETSc team. My responsibilities here are working on the backend but what I&#8217;m most excited about is improving Mercurial support. I&#8217;ve been a contrib
]]>
<![CDATA[
utor to Mercurial for about three years and have loved Mercurial for its human-friendly user interface and supportive community. Some of my work for Mercurial includes the addition of namespaces and improving the in-memory context object. I also wrote the experimental remotenames extension (to be added to core Mercurial in the future) and the most fun smb extension. The future of Mercurial In my time with Mercurial, I have seen it grow in fascinating ways. These include the concept of changeset evolution coming to life and the announcement of Facebook choosing Mercurial over Git. The future of Mercurial is that of scalability and because of that, I believe the best days of Mercurial are ahead. Mercurial 3.4 Sprint at PyCon Here at Bitbucket, with the support of the team, I&#8217;d like to build up relations with the community and the first step was attending the Mercurial 3.4 Sprint at PyCon in Montreal this year. There were many topics discussed: new manifest version, narrow checkouts, commit signing, obsolescence marker exchange, evolve UI, reflog equivalent, new path options, and remote bookmarks. Of these topics, manifests received the most attention. A new manifest format means we can enable checkouts of only a subdirectory (called a &#8216;narrow&#8217; checkout). On top of that, it opens the door for only retrieving a truncated history (called a &#8216;shallow&#8217; checkout). My personal favorites are the inclusion of remote bookmarks that will be built on top of the journal i.e. reflog equivalent, and the new path features i.e. the ability to define a custom revset for pushing to a particular path. Here are some pictures from the event (photos by Ryan McElroy and me): Celebrating Mercurial&#8217;s 10th anniversary Now that I am here, with the support of Atlassian and the Bitbucket team, we will be improving relations with the Mercurial community by hosting many events this year. We begin with celebrating Mercurial&#8217;s 10th anniversary on Wed, May 6th by hosting the first official Bay Area Mercurial Meetup at Atlassian&#8217;s San Francisco office. We will have speakers from Google to talk about scaling Mercurial and Facebook to talk about why they picked Mercurial as the version control system and what this means for the future of Mercurial. Register immediately if you are interested since we can only accommodate 100 attendees. We look forward to seeing you next week with the rest of the Mercurial community. [*Note: An earlier version had accidentally stated that Google had chosen Mercurial over Git but that is untrue. Google is working on Mercurial scalability, but also uses Git and contributes to Git development.]
]]>
</description>
<content:encoded>
<![CDATA[
<p>Hi, I&#8217;m Sean, the newest member on the Bitbucket team. I currently work on <a class="external-link" href="http://mercurial.selenic.com/" target="_blank" rel="nofollow">Mercurial</a> but come from a scientific background with the <a class="external-link" href="http://www.mcs.anl.gov/petsc" target="_blank" rel="nofollow">PETSc</a> team. My responsibilities here are working on the backend but what I&#8217;m most excited about is improving Mercurial support.</p> <p>I&#8217;ve been a contributor to Mercurial for <span class="inline-comment-marker">about three years</span> and have loved Mercurial for its human-friendly user interface and supportive community. Some of my work for Mercurial includes the addition of namespaces and improving the in-memory context object. I also wrote the experimental <a class="external-link" href="https://bitbucket.org/seanfarley/hgremotenames" target="_blank" rel="nofollow">remotenames extension</a> (to be added to core Mercurial in the future) and the most fun <a class="external-link" href="http://youtu.be/qtAuQXeQrMQ" target="_blank" rel="nofollow">smb extension</a>.</p> <p><strong>The future of Mercurial</strong></p> <p>In my time with Mercurial, I have seen it grow in fascinating ways. These include the concept of changeset evolution coming to life and the announcement of Facebook choosing Mercurial over Git. The future of Mercurial is that of scalability and because of that, <span class="inline-comment-marker">I believe the best days of Mercurial are ahead</span>.</p> <p><strong>Mercurial 3.4 Sprint at PyCon</strong></p> <p>Here at Bitbucket, with the support of the team, I&#8217;d like to build up relations with the community and the first step was attending the <a class="external-link" href="http://mercurial.selenic.com/wiki/3.4sprint" target="_blank" rel="nofollow">Mercurial 3.4 Sprint</a> at PyCon in Montreal this year. There were many topics discussed: new manifest version, narrow checkouts, commit signing, obsolescence marker exchange, evolve UI, reflog equivalent, new path options, and remote bookmarks. Of these topics, manifests received the most attention.