Fix L4 nesting for cxx"""
Originally this was just supposed to be a bug-fix for the icxx"""
version of this functionality, but I got a bit carried away.
This reuses the icxx""" functionality for the cxx""" one, which
is a little awkward, since they didn't share any code before even
though they did very similar things.