Fix #149
by properly throwing an error if implicit return type dedcution fails rather
than just crashing. It would be possible to intercept this specific case at
various stages to actually return the function pointer, but I'd rather have
a clear explanation of what the return type deduction does. Eventually it
might be nice to have the return type deduction do function-to-pointer decay,
but I don't think that's possible without hacking clang, so is not a priority
for now.