Allow Cxx to load julia.h
Cxx's boot.h comes with some minimal definitions from julia.h, to allow it
to call back into the julia runtime. However, this also prevents it from
loading julia.h due to conflicting type names. Fix this by optionally delaying
loading boot.h until the user had a chance to run julia.h. This isn't an ideal
solution. Maybe in the future some tricks can be played with symbol aliasing.
But for now, this is a simple solution that works.