Fix #95
Not really happy with this solution, but passing it through an argument should ensure that
it gets rooted at least through the lifetime of the call. Of course, the GC optimizations
may get better in the future and make this unsafe again. At that point we'll need to
revisit this.