Fix #209
This is a two-pronged approach. First, we ignore any trailing null
statements when computing implicit return. This is similar to what
julia does, but is contrary to GNU statement expressions, so Clang
will still diagnose the unsed result. As a result, strip any trailing
semicolon in the REPL as well.