Provide override for standard headers in C++
Using Cxx on a system like Gentoo or some supercomputers with multiple
compilers will result in Cxx not finding the standard library.  This
change allows users to specify where to find the system includes in a
way that doesn't requiring forking Cxx to make these changes.  Since
these systems are inherently non-standard, these changes are unlikely to
be fixed with a change to `CollectLinuxHeaderPaths!` so providing an
override is better.